Fix metrics collection for privileged process on Windows (#104)
## What does this PR do?

It improves metric collection on Windows hosts so we can collect some
metrics from privileged process like Elastic Endpoint or Elastic-Agent.
In order to achieve that metrics collection for Windows are grouped by
the access level required to collect metrics. First the metrics that can
be collected with `PROCESS_QUERY_LIMITED_INFORMATION` are collected,
then the others are collected. In case the second batch fails, we still
report the partial metrics.

## Why is it important?

It allows us to collect CPU and memory metrics from Endpoint and
Elastic-Agent, which improves our monitoring dashboards.

## Related issues

- Closes https://github.com/elastic/beats/issues/17314